Use Icon column to render radio with icons
